fix(routines): split power-saving tests out of service_trigger_tests.rs
service_trigger_tests.rs exceeded the 700-line linecheck gate once the
merge landed the power-saving trigger/lock tests alongside main's own
trigger-test growth. Move them to a dedicated service_power_saving_tests.rs,
matching the existing one-file-per-concern test layout.
