+++ src/routines/service_trigger_tests.rs
@@ -562,119 +562,6 @@ fn svc_trigger_scheduled_returns_locked_when_globally_locked() {
     );
 }
 
-#[test]
-fn svc_trigger_returns_locked_when_disabled() {
-    let _home = TempHome::set();
-    let store = new_store();
-    let mut routine = make_routine("disabled-trig-id", "Disabled Trigger Test ZZZ", 1, 1);
-    routine.enabled = false;
-    store
-        .lock()
-        .unwrap()
-        .insert("disabled-trig-id".into(), routine);
-
-    let result = svc_trigger(&store, "disabled-trig-id");
-    assert!(
-        matches!(result, Err(AppError::Locked(ref msg)) if msg.contains("disabled")),
-        "expected a Locked error naming disabled, got {result:?}"
-    );
-}
-
-#[test]
-fn svc_trigger_returns_locked_when_power_saving() {
-    let _home = TempHome::set();
-    let store = new_store();
-    let mut routine = make_routine(
-        "power-saving-trig-id",
-        "Power Saving Trigger Test ZZZ",
-        1,
-        1,
-    );
-    routine.power_saving = true;
-    store
-        .lock()
-        .unwrap()
-        .insert("power-saving-trig-id".into(), routine);
-
-    let result = svc_trigger(&store, "power-saving-trig-id");
-    assert!(
-        matches!(result, Err(AppError::Locked(ref msg)) if msg.contains("power-saving")),
-        "expected a Locked error naming power-saving, got {result:?}"
-    );
-}
-
-#[test]
-fn svc_trigger_scheduled_returns_locked_when_disabled() {
-    let _home = TempHome::set();
-    let store = new_store();
-    let mut routine = make_routine("disabled-sched-id", "Disabled Sched Test ZZZ", 1, 1);
-    routine.enabled = false;
-    store
-        .lock()
-        .unwrap()
-        .insert("disabled-sched-id".into(), routine);
-
-    let result = svc_trigger_scheduled(&store, "disabled-sched-id");
-    assert!(
-        matches!(result, Err(AppError::Locked(ref msg)) if msg.contains("disabled")),
-        "expected a Locked error naming disabled, got {result:?}"
-    );
-}
-
-#[test]
-fn svc_trigger_scheduled_returns_locked_when_power_saving() {
-    let _home = TempHome::set();
-    let store = new_store();
-    let mut routine = make_routine("power-saving-sched-id", "Power Saving Sched Test ZZZ", 1, 1);
-    routine.power_saving = true;
-    store
-        .lock()
-        .unwrap()
-        .insert("power-saving-sched-id".into(), routine);
-
-    let result = svc_trigger_scheduled(&store, "power-saving-sched-id");
-    assert!(
-        matches!(result, Err(AppError::Locked(ref msg)) if msg.contains("power-saving")),
-        "expected a Locked error naming power-saving, got {result:?}"
-    );
-}
-
-#[test]
-fn svc_set_power_saving_missing_routine_not_found() {
-    let _home = TempHome::set();
-    assert!(matches!(
-        svc_set_power_saving(&new_store(), "nope", true),
-        Err(AppError::NotFound)
-    ));
-}
-
-#[test]
-fn svc_set_power_saving_sets_and_clears_without_touching_enabled() {
-    let _home = TempHome::set();
-    let store = new_store();
-    let routine = make_routine("power-saving-set-id", "Power Saving Set Test ZZZ", 1, 1);
-    store
-        .lock()
-        .unwrap()
-        .insert("power-saving-set-id".into(), routine);
-
-    let updated = svc_set_power_saving(&store, "power-saving-set-id", true).unwrap();
-    assert!(updated.power_saving);
-    assert!(updated.enabled, "enabled must be untouched");
-    assert!(
-        store
-            .lock()
-            .unwrap()
-            .get("power-saving-set-id")
-            .unwrap()
-            .power_saving
-    );
-
-    let updated = svc_set_power_saving(&store, "power-saving-set-id", false).unwrap();
-    assert!(!updated.power_saving);
-    assert!(updated.enabled);
-}
-
 #[cfg(unix)]
 #[test]
 fn svc_trigger_returns_internal_on_write_failure() {
